Added function to reset the GUI between videos
A new function resets the GUI when a new video is opened. This
clears e.g. the status bar info, etc, and is especially useful to
leave a tidy UI even if opening a new video file fails (tipycally
because e.g. a missing tifffile module).

A pop-up window was also added to display an error when a request to
open a video file fails due to a missing module.