Save data in long table format
ROI intensity data can now be saved in 'long' or 'wide' table format.
It is currently hardcoded to be saved  in long format as default
because this is the most convenient of the two for me. Future versions
will allow the user to choose in the UI which of these two formats to
use.